fix warnings
fix WS data received at invalid state

+++ test/client/WsClient.cpp
@@ -94,7 +94,7 @@ void WsClient::onData(KMBuffer &buf)
         timer_.cancel();
         std::chrono::steady_clock::time_point end_point = std::chrono::steady_clock::now();
         std::chrono::milliseconds diff_ms = std::chrono::duration_cast<std::chrono::milliseconds>(end_point - start_point_);
-        printf("spent %lld ms to echo %u packets\n", diff_ms.count(), max_send_count_);
+        printf("spent %lld ms to echo %u packets\n", (long long int)diff_ms.count(), max_send_count_);
         loop_->removeObject(conn_id_);
     }
 }
